python_api/version_1/requirements.txt
Flask
requests
xmltodict
pycryptodome
